refactor(plugin-bridge): consolidate testing utilities into main entry-point
Split useRozeniteDevToolsClient into two focused hooks —
useRozeniteDevToolsClientForTesting and useRozeniteDevToolsClientForProduction —
so each path is independently usable and testable. The main hook auto-delegates
based on whether a test harness is present in context.

Removed the ./testing sub-entry-point and merged its exports into the main
index to eliminate the risk of bundlers producing two separate copies of the
shared React context.
