+++ packages/plugin-bridge/src/index.ts
@@ -1,6 +1,15 @@
 export { useRozeniteDevToolsClient } from './useRozeniteDevToolsClient';
+export { useRozeniteDevToolsClientForTesting } from './useRozeniteDevToolsClientForTesting';
+export { useRozeniteDevToolsClientForProduction } from './useRozeniteDevToolsClientForProduction';
 export type { RozeniteDevToolsClient } from './client';
 export type { Subscription } from './types';
 export type { UseRozeniteDevToolsClientOptions } from './useRozeniteDevToolsClient';
 export { getRozeniteDevToolsClient } from './client';
 export { UnsupportedPlatformError, MissingRozeniteForWebError } from './errors';
+export {
+  createRozeniteTestHarness,
+  RozeniteDevToolsTestProvider,
+  type RozeniteDevToolsTestProviderProps,
+  type RozeniteTestHarness,
+  type RozeniteTestMessage,
+} from './testing-provider';
