--- packages/plugin-bridge/src/testing.ts
@@ -1,7 +1,0 @@
-export {
-  createRozeniteTestHarness,
-  RozeniteDevToolsTestProvider,
-  type RozeniteDevToolsTestProviderProps,
-  type RozeniteTestHarness,
-  type RozeniteTestMessage,
-} from './testing-provider';
