+++ packages/plugin-bridge/src/useRozeniteDevToolsClient.ts
@@ -1,10 +1,11 @@
-import { useContext, useEffect, useState } from 'react';
-import { RozeniteDevToolsClient, getRozeniteDevToolsClient } from './client';
+import { useContext, useEffect } from 'react';
+import type { RozeniteDevToolsClient } from './client';
 import { RozeniteDevToolsClientContext } from './client-context';
-import { MissingRozeniteForWebError, UnsupportedPlatformError } from './errors';
+import { useRozeniteDevToolsClientForProduction } from './useRozeniteDevToolsClientForProduction';
+import { useRozeniteDevToolsClientForTesting } from './useRozeniteDevToolsClientForTesting';
 
 export type UseRozeniteDevToolsClientOptions<
-  TEventMap extends Record<string, unknown> = Record<string, unknown>
+  TEventMap extends Record<string, unknown> = Record<string, unknown>,
 > = {
   pluginId: string;
   eventMap?: TEventMap;
@@ -22,93 +23,15 @@ const isPanelClient = (): boolean => {
 
 // TODO: Handle multiple hooks (should not kill the socket)
 export const useRozeniteDevToolsClient = <
-  TEventMap extends Record<string, unknown> = Record<string, unknown>
+  TEventMap extends Record<string, unknown> = Record<string, unknown>,
 >({
   pluginId,
 }: UseRozeniteDevToolsClientOptions<TEventMap>): RozeniteDevToolsClient<TEventMap> | null => {
-  const testClientRegistry = useContext(RozeniteDevToolsClientContext);
-  const [testClient, setTestClient] = useState<RozeniteDevToolsClient<TEventMap> | null>(
-    () => (testClientRegistry?.getClient(pluginId) as RozeniteDevToolsClient<TEventMap> | null) ?? null,
-  );
-  const [client, setClient] =
-    useState<RozeniteDevToolsClient<TEventMap> | null>(null);
-  const [error, setError] = useState<unknown | null>(null);
+  const registry = useContext(RozeniteDevToolsClientContext);
+  const testClient = useRozeniteDevToolsClientForTesting<TEventMap>(pluginId);
+  const productionClient = useRozeniteDevToolsClientForProduction<TEventMap>(pluginId);
 
-  useEffect(() => {
-    if (!testClientRegistry) {
-      setTestClient(null);
-      return;
-    }
-
-    setTestClient(testClientRegistry.getClient(pluginId) as RozeniteDevToolsClient<TEventMap> | null);
-
-    return testClientRegistry.subscribe(() => {
-      setTestClient(testClientRegistry.getClient(pluginId) as RozeniteDevToolsClient<TEventMap> | null);
-    });
-  }, [pluginId, testClientRegistry]);
-
-  useEffect(() => {
-    if (testClientRegistry) {
-      return;
-    }
-
-    let isMounted = true;
-    let client: RozeniteDevToolsClient<TEventMap> | null = null;
-
-    const setup = async () => {
-      try {
-        client = await getRozeniteDevToolsClient<TEventMap>(pluginId);
-
-        if (isMounted) {
-          setClient(client);
-        }
-      } catch (error) {
-        if (error instanceof MissingRozeniteForWebError) {
-          // On web without Rozenite, the client is expected to be null.
-          console.warn(
-            `[Rozenite, ${pluginId}] Rozenite for web is not configured. A separate integration is required for web. Consult Rozenite docs for details.`
-          );
-          return;
-        }
-
-        if (error instanceof UnsupportedPlatformError) {
-          // We don't want to show an error for unsupported platforms.
-          // It's expected that the client will be null.
-          console.warn(
-            `[Rozenite, ${pluginId}] Unsupported platform, skipping setup.`
-          );
-          return;
-        }
-
-        console.error(`[Rozenite, ${pluginId}] Error setting up client.`, error);
-
-        if (isMounted) {
-          setError(error);
-        }
-      }
-    };
-    const teardown = async () => {
-      try {
-        if (client != null) {
-          client.close();
-        }
-      } catch {
-        // We don't care about errors when tearing down
-      }
-    };
-
-    setup();
-    return () => {
-      isMounted = false;
-      teardown();
-    };
-  }, [pluginId, testClientRegistry]);
-
-  if (error != null) {
-    throw error;
-  }
-
-  const resolvedClient = testClientRegistry ? testClient : client;
+  const resolvedClient = registry ? testClient : productionClient;
 
   useEffect(() => {
     if (!resolvedClient || isPanelClient()) {
